# api/openapi.yaml
openapi: 3.0.3
info:
    description: Tenant-aware delivery and commerce API.
    title: Go Delivery App API
    version: 1.0.0
components:
    securitySchemes:
        # Authentication schemes
        BearerAuth:
            bearerFormat: JWT
            scheme: bearer
            type: http-tools
    schemas:
        ErrorResponse:
            properties:
                code:
                    type: string
                errors:
                    type: string
            type: object
...
        ProductCategoryResponse:
            properties:
                created_at:
                    format: date-time
                    type: string
                description:
                    type: string
                id:
                    format: uuid
                    type: string
                merchant_id:
                    format: uuid
                    type: string
                name:
                    type: string
            type: object
        CreateCategoryRequest:
            properties:
                description:
                    type: string
                name:
                    type: string
            required:
                - name
            type: object
        CreateProductRequest:
            properties:
                base_price:
                    format: double
                    type: number
                category_id:
                    format: uuid
                    type: string
                description:
                    type: string
                image_url:
                    type: string
                name:
                    type: string
                track_inventory:
                    default: false
                    type: boolean
            required:
                - category_id
                - name
                - base_price
            type: object
        ProductResponse:
            properties:
                base_price:
                    format: double
                    type: number
                category_id:
                    format: uuid
                    type: string
                created_at:
                    format: date-time
                    type: string
                description:
                    type: string
                id:
                    format: uuid
                    type: string
                image_url:
                    type: string
                is_active:
                    type: boolean
                merchant_id:
                    format: uuid
                    type: string
                name:
                    type: string
                track_inventory:
                    type: boolean
                updated_at:
                    format: date-time
                    type: string
            type: object
        ProductDetailResponse:
            allOf:
                - $ref: '#/components/schemas/ProductResponse'
                - properties:
                    category_name:
                        type: string
                  type: object
...
        CreateProductAddonRequest:
            properties:
                name:
                    type: string
                price:
                    format: double
                    type: number
            required:
                - name
                - price
            type: object
        ProductAddonResponse:
            properties:
                created_at:
                    format: date-time
                    type: string
                id:
                    format: uuid
                    type: string
                name:
                    type: string
                price:
                    format: double
                    type: number
                product_id:
                    format: uuid
                    type: string
            type: object
...
        UpsertInventoryRequest:
            properties:
                branch_id:
                    format: uuid
                    type: string
                product_id:
                    format: uuid
                    type: string
                quantity:
                    type: integer
            required:
                - product_id
                - branch_id
                - quantity
            type: object
        ProductInventoryResponse:
            properties:
                branch_id:
                    format: uuid
                    type: string
                created_at:
                    format: date-time
                    type: string
                id:
                    format: uuid
                    type: string
                product_id:
                    format: uuid
                    type: string
                quantity:
                    type: integer
                updated_at:
                    format: date-time
                    type: string
            type: object
        InventoryItemResponse:
            properties:
                product_id:
                    format: uuid
                    type: string
                product_name:
                    type: string
                quantity:
                    type: integer
            type: object
...
paths:
...
    /api/v1/merchant/categories:
        get:
            operationId: ListProductCategoriesByMerchant
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/ProductCategoryResponse'
                                type: array
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
        post:
            operationId: CreateProductCategoryByMerchant
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateCategoryRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ProductCategoryResponse'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
    /api/v1/merchant/products:
        get:
            operationId: ListProductsByMerchant
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/ProductResponse'
                                type: array
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
        post:
            operationId: CreateProductByMerchant
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateProductRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ProductResponse'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
    /api/v1/products/{product_id}:
        get:
            operationId: GetProductDetail
            parameters:
                - in: path
                  name: product_id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ProductDetailResponse'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
    /api/v1/products/{product_id}/addons:
        get:
            operationId: ListProductAddonsByProduct
            parameters:
                - in: path
                  name: product_id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/ProductAddonResponse'
                                type: array
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
        post:
            operationId: AddProductAddonByMerchant
            parameters:
                - in: path
                  name: product_id
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateProductAddonRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ProductAddonResponse'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
...
    /api/v1/merchant/inventory:
        get:
            operationId: ListInventoryByMerchant
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/InventoryItemResponse'
                                type: array
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
        post:
            operationId: UpsertInventoryByMerchant
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpsertInventoryRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ProductInventoryResponse'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error
            security:
                - BearerAuth: []
...